requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==3.1.3
django-bootstrap4==2.3.1
django-classy-tags==2.0.0
django-cms==3.8.0
django-colorfield==0.3.2
django-cors-middleware==1.5.0
django-extensions==3.0.9
django-filer==2.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-registration==3.1.1
django-sekizai==2.0.0
django-treebeard==4.3.1
djangocms-admin-style==2.0.1
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-snippet==3.0.0
djangocms-style==3.0.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==4.0.0
djangocms-transfer==1.0.0
djangocms-video==3.0.0
djangorestframework==3.12.2
easy-thumbnails==2.7
html5lib==1.1
idna==2.10
importlib-metadata==2.0.0
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
Unidecode==1.1.1
urllib3==1.26.2
webencodings==0.5.1
zipp==3.4.0
